+++ yak-security/src/main/java/io/yak/framework/security/common/dto/role/RoleAssignDTO.java
@@ -2,19 +2,19 @@ package io.yak.framework.security.common.dto.role;
 
 import java.util.List;
 public class RoleAssignDTO {
-  private Integer id;
-  private List<Integer> idList;
+  private Long id;
+  private List<Long> idList;
   private Boolean flag;
 
-  public Integer getId() { return this.id; }
+  public Long getId() { return this.id; }
 
-  public List<Integer> getIdList() { return this.idList; }
+  public List<Long> getIdList() { return this.idList; }
 
   public Boolean getFlag() { return this.flag; }
 
-  public void setId(Integer id) { this.id = id; }
+  public void setId(Long id) { this.id = id; }
 
-  public void setIdList(List<Integer> idList) { this.idList = idList; }
+  public void setIdList(List<Long> idList) { this.idList = idList; }
 
   public void setFlag(Boolean flag) { this.flag = flag; }
 
@@ -29,8 +29,8 @@ public class RoleAssignDTO {
     if (!other.canEqual(this)) {
       return false;
     }
-    Integer this$id = this.getId();
-    Integer other$id = other.getId();
+    Long this$id = this.getId();
+    Long other$id = other.getId();
     if (this$id == null ? other$id != null
                         : !((Object)this$id).equals(other$id)) {
       return false;
@@ -41,8 +41,8 @@ public class RoleAssignDTO {
                           : !((Object)this$flag).equals(other$flag)) {
       return false;
     }
-    List<Integer> this$idList = this.getIdList();
-    List<Integer> other$idList = other.getIdList();
+    List<Long> this$idList = this.getIdList();
+    List<Long> other$idList = other.getIdList();
     return !(this$idList == null ? other$idList != null
                                  : !((Object)this$idList).equals(other$idList));
   }
@@ -54,11 +54,11 @@ public class RoleAssignDTO {
   public int hashCode() {
     int PRIME = 59;
     int result = 1;
-    Integer $id = this.getId();
+    Long $id = this.getId();
     result = result * 59 + ($id == null ? 43 : ((Object)$id).hashCode());
     Boolean $flag = this.getFlag();
     result = result * 59 + ($flag == null ? 43 : ((Object)$flag).hashCode());
-    List<Integer> $idList = this.getIdList();
+    List<Long> $idList = this.getIdList();
     result =
         result * 59 + ($idList == null ? 43 : ((Object)$idList).hashCode());
     return result;
