+++ yak-security/src/main/java/io/yak/framework/security/common/dto/role/RoleSaveDTO.java
@@ -2,20 +2,20 @@ package io.yak.framework.security.common.dto.role;
 
 import java.util.List;
 public class RoleSaveDTO {
-  private Integer id;
+  private Long id;
   private String roleName;
   private String description;
-  private List<Integer> permissionIdList;
+  private List<Long> permissionIdList;
 
-  public Integer getId() { return this.id; }
+  public Long getId() { return this.id; }
 
   public String getRoleName() { return this.roleName; }
 
   public String getDescription() { return this.description; }
 
-  public List<Integer> getPermissionIdList() { return this.permissionIdList; }
+  public List<Long> getPermissionIdList() { return this.permissionIdList; }
 
-  public void setId(Integer id) { this.id = id; }
+  public void setId(Long id) { this.id = id; }
 
   public void setRoleName(String roleName) { this.roleName = roleName; }
 
@@ -23,7 +23,7 @@ public class RoleSaveDTO {
     this.description = description;
   }
 
-  public void setPermissionIdList(List<Integer> permissionIdList) {
+  public void setPermissionIdList(List<Long> permissionIdList) {
     this.permissionIdList = permissionIdList;
   }
 
@@ -38,8 +38,8 @@ public class RoleSaveDTO {
     if (!other.canEqual(this)) {
       return false;
     }
-    Integer this$id = this.getId();
-    Integer other$id = other.getId();
+    Long this$id = this.getId();
+    Long other$id = other.getId();
     if (this$id == null ? other$id != null
                         : !((Object)this$id).equals(other$id)) {
       return false;
@@ -57,8 +57,8 @@ public class RoleSaveDTO {
             : !this$description.equals(other$description)) {
       return false;
     }
-    List<Integer> this$permissionIdList = this.getPermissionIdList();
-    List<Integer> other$permissionIdList = other.getPermissionIdList();
+    List<Long> this$permissionIdList = this.getPermissionIdList();
+    List<Long> other$permissionIdList = other.getPermissionIdList();
     return !(
         this$permissionIdList == null
             ? other$permissionIdList != null
@@ -72,14 +72,14 @@ public class RoleSaveDTO {
   public int hashCode() {
     int PRIME = 59;
     int result = 1;
-    Integer $id = this.getId();
+    Long $id = this.getId();
     result = result * 59 + ($id == null ? 43 : ((Object)$id).hashCode());
     String $roleName = this.getRoleName();
     result = result * 59 + ($roleName == null ? 43 : $roleName.hashCode());
     String $description = this.getDescription();
     result =
         result * 59 + ($description == null ? 43 : $description.hashCode());
-    List<Integer> $permissionIdList = this.getPermissionIdList();
+    List<Long> $permissionIdList = this.getPermissionIdList();
     result = result * 59 + ($permissionIdList == null
                                 ? 43
                                 : ((Object)$permissionIdList).hashCode());
